Moved HeatMap plots to their own module
