+++ holoviews/plotting/mpl/__init__.py
@@ -14,6 +14,7 @@ from .annotation import * # noqa (API import)
 from .chart import * # noqa (API import)
 from .chart3d import * # noqa (API import)
 from .graphs import * # noqa (API import)
+from .heatmap import * # noqa (API import)
 from .path import * # noqa (API import)
 from .plot import * # noqa (API import)
 from .raster import * # noqa (API import)
